Add implementation of FC-DenseNet103 for SimCLR
Added FCDenseNet103Encoder class which contains the downsampling path of
a normal network and a projection head – global pool to 2x2 followed by
a MLP with 1 hidden layer and ReLU. This model will be trained in SimCLR
framework in later experiments.

Also removed some old comments from the definition of FC-DenseNet103 and
renamed some tests to make them fit on the screen in verbose pytest
mode.